Refactor schema to follow new folder structure
