+++ src/schema/organization/teams/index.js
@@ -1,0 +1,19 @@
+import {
+  GraphQLInt,
+  GraphQLList,
+  GraphQLNonNull,
+} from 'graphql';
+
+import TeamType from './TeamType';
+import resolvers from './resolvers';
+
+export default {
+  type: new GraphQLList(TeamType),
+  args: {
+    maxNumberOfTeams: {
+      type: new GraphQLNonNull(GraphQLInt),
+      description: 'Max number of results from this query'
+    },
+  },
+  resolve: resolvers.Query.teams,
+}
