+++ src/schema/user/queries.js
@@ -1,0 +1,11 @@
+import UserType from './UserType';
+import resolvers from './resolvers';
+
+const me = {
+  type: UserType,
+  resolve: resolvers.Query.me,
+}
+
+export default {
+  me,
+}
